pyproject.toml
[tool.poetry]
name = "101-regex"
version = "0.1.0"
description = "Projeto para aula de extração de dados em PDF e uso de Regex"
authors = ["Fabio Cantarim <fabio.cantarim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
camelot-py = "^0.11.0"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.3"
ghostscript = "^0.7"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
unidecode = "^1.3.8"
pypdf2 = "^3.0.1"
black = "^24.8.0"
isort = "^5.13.2"
schedule = "^1.2.2"
pytz = "^2024.2"
boto3 = "^1.35.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
